This MVVM is a bit of a headache, but it's coming together.

+++ DataAccessLayer/SqlTagM.cs
@@ -35,6 +35,12 @@ namespace AI_Note_Review
 
         public void RemoveTagRegEx(SqlTagRegExVM str)
         {
+            MessageBoxResult mr = MessageBox.Show("Are you sure you want to delete this search term from the database? This is permenant.", "Confirm Delete", MessageBoxButton.YesNo);
+            if (mr != MessageBoxResult.Yes)
+            {
+                return;
+            }
+
             string sql = $"Delete from TagRegEx WHERE TagRegExID={str.TagRegExID};";
             using (IDbConnection cnn = new SQLiteConnection("Data Source=" + SqlLiteDataAccess.SQLiteDBLocation))
             {
